refs 126149: rename LdapGroupSyncService filename
